+++ src/Models/Spell.php
@@ -62,6 +62,6 @@ class Spell extends AbstractModel
     }
 
     public function isHarmful() {
-        return $this->category == 'hostile' || $this->key == 'burning';
+        return $this->category == 'hostile' || $this->category == 'war' || $this->key == 'burning';
     }
 }